requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.3
ffmpy==0.4.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==5.4.0
gradio_client==1.4.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.137
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
ruff==0.7.1
safehttpx==0.1.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
yarl==1.16.0